moved MovieList to Cine folder and edited all corresponding path
